requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
billiard==4.1.0
biopython==1.81
biotite==0.37.0
bokeh
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
cmake==3.26.4
comm==0.1.3
contourpy
cycler==0.11.0
datasets
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
executing==1.2.0
fair-esm==2.0.0
fastjsonschema==2.17.1
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
googledrivedownloader==0.4
huggingface-hub==0.16.2
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lit==16.0.6
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib
matplotlib-inline
mmtf-python==1.1.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.1
numba
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
rdflib==6.3.2
regex==2023.6.3
requests==2.31.0
responses==0.23.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-learn==1.3.0
scipy
seaborn==0.12.2
sequence-models==1.6.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
#torch-cluster==1.6.1
#torch-geometric==2.3.1
#torch-scatter==2.1.1
#torch-sparse==0.6.17
#torch-spline-conv==1.2.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
types-PyYAML==6.0.12.10
types-toml==0.10.8.6
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
xxhash==3.2.0
xyzservices==2023.5.0
yacs==0.1.8
yarl==1.9.2
